requirements.txt
absl-py==0.2.2
appnope==0.1.0
asn1crypto==0.23.0
astor==0.7.1
backcall==0.1.0
bcolz==1.2.1
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.2
chardet==3.0.4
conda==4.4.7
cryptography==2.1.4
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.3.0
dill==0.2.7.1
entrypoints==0.2.3
fastai==0.7.0
feather-format==0.4.0
ftfy==4.4.3
gast==0.2.0
graphviz==0.8.4
grpcio==1.13.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
isoweek==1.3.3
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mizani==0.4.6
Morfessor==2.0.3
msgpack-numpy==0.4.1
msgpack-python==0.4.8
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
neo4j-driver==1.6.1
neotime==1.0.0
networkx==2.1
networkx-neo4j==0.0.1
nltk==3.2.5
notebook==5.4.1
numpy==1.14.3
opencv-python==3.4.1.15
palettable==3.1.1
pandas==0.21.0
pandas-summary==0.0.41
pandocfilters==1.4.2
parso==0.2.0
pathlib==1.0.1
patsy==0.5.0
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.2.0
pipdeptree==0.13.0
plac==0.9.6
plotnine==0.3.0
polyglot==16.7.4
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.5.2
pyarrow==0.9.0.post1
pycld2==0.31
pycosat==0.6.3
pycparser==2.18
Pygments==2.2.0
PyICU==1.9.8
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
regex==2017.4.5
requests==2.18.4
reverse-geocoder==1.5.1
ruamel-yaml==0.11.14
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
sklearn-neo4j==0.0.1
sklearn-pandas==1.6.0
spacy==2.0.3
statsmodels==0.9.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.1
toolz==0.8.2
torch==0.3.1
torchtext==0.2.3
torchvision==0.2.1
tornado==5.0.2
tqdm==4.19.4
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
